requirements.txt
django == 4.0.4
pillow

django-admin-interface
django-crispy-forms
djangorestframework

mysqlclient
pymysql